+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Entity descriptor counters
 - Detection of arriving and departing network interfaces (and link status)
 - [Sort connection matrix by EntityID](https://github.com/christophe-calmejane/Hive/issues/59)
+- Connection matrix complete overhaul (optimization, smart connections)
 
 ### Changed
 - Only displaying _Ethernet_ kind interfaces